Fix site_id when no website is existing yet
The existing workaround to New-WebSite bug forces the site ID to 1 when
no website exists, that overrides the site_id argument specified by the
user. A try/catch block is added to use the workaround when the call to
New-WebSite fails, typically when no site ID is specified and no website
is existing yet.